+++ packages/preset/src/index.js
@@ -3,5 +3,4 @@ export {
   API_ROOT,
   MAX_BODY_BYTES,
   createApiHandler,
-  installServerRoutes,
 } from './server.js'